Update asset URLs for webpack imported CSS
Unlike the regular CSS inside used in the components addons, these CSS files are included through javascript import statements that webpack figures out.

The important (and confusing) aspect of this is that the url is relative for the location of the files once embroider is done it's stage 1 compat build. This stage 1 build moves a lot of things around.

You'll receive an error like :
```
Can't resolve '../@cardstack/boxel/images/icons/check-mark.svg' in '$TMPDIR/embroider/095181/tests/dummy/css/components'
```
`$TMPDIR/embroider/095181/tests/dummy/css/components` is the location of the result of the stage 1 embroider build. The import statement that is failing is in that directory. You'll need to open that directory, navigate around to find where the file that you're trying to reference in css has ended up and adjust your relative path accordingly.